Move LowPriorityImplicits to a com.softwaremill.quicklens to avoid conflicts with other softwaremill libraries
